+++ pollux/lib/core/src/main/scala/io/iohk/atala/pollux/core/model/error/CredentialServiceError.scala
@@ -2,14 +2,15 @@ package io.iohk.atala.pollux.core.model.error
 
 import java.util.UUID
 import io.iohk.atala.pollux.vc.jwt.W3cCredentialPayload
+import io.iohk.atala.pollux.core.model.DidCommID
 
 sealed trait CredentialServiceError
 
 object CredentialServiceError {
   final case class RepositoryError(cause: Throwable) extends CredentialServiceError
-  final case class RecordIdNotFound(recordId: UUID) extends CredentialServiceError
-  final case class OperationNotExecuted(recordId: UUID, info: String) extends CredentialServiceError
-  final case class ThreadIdNotFound(thid: UUID) extends CredentialServiceError
+  final case class RecordIdNotFound(recordId: DidCommID) extends CredentialServiceError
+  final case class OperationNotExecuted(recordId: DidCommID, info: String) extends CredentialServiceError
+  final case class ThreadIdNotFound(thid: DidCommID) extends CredentialServiceError
   final case class InvalidFlowStateError(msg: String) extends CredentialServiceError
   final case class UnexpectedError(msg: String) extends CredentialServiceError
   final case class UnsupportedDidFormat(did: String) extends CredentialServiceError