+++ pollux/lib/core/src/main/scala/io/iohk/atala/pollux/core/repository/CredentialRepositoryInMemory.scala
@@ -2,10 +2,9 @@ package io.iohk.atala.pollux.core.repository
 
 import io.iohk.atala.mercury.protocol.issuecredential.IssueCredential
 import io.iohk.atala.mercury.protocol.issuecredential.RequestCredential
-import io.iohk.atala.pollux.core.model.IssueCredentialRecord
+import io.iohk.atala.pollux.core.model._
 import io.iohk.atala.pollux.core.model.IssueCredentialRecord.ProtocolState
 import io.iohk.atala.pollux.core.model.IssueCredentialRecord.PublicationState
-import io.iohk.atala.pollux.core.model.ValidIssuedCredentialRecord
 import io.iohk.atala.pollux.core.model.error.CredentialRepositoryError._
 import io.iohk.atala.prism.crypto.MerkleInclusionProof
 import zio.*
@@ -14,18 +13,18 @@ import java.time.Instant
 import java.util.UUID
 
 class CredentialRepositoryInMemory(
-    storeRef: Ref[Map[UUID, IssueCredentialRecord]],
+    storeRef: Ref[Map[DidCommID, IssueCredentialRecord]],
     maxRetries: Int
 ) extends CredentialRepository[Task] {
 
   override def updateCredentialRecordPublicationState(
-      recordId: UUID,
+      recordId: DidCommID,
       from: Option[PublicationState],
       to: Option[PublicationState]
   ): Task[Int] = {
     for {
       store <- storeRef.get
-      maybeRecord = store.find((uuid, record) => uuid == recordId && record.publicationState == from).map(_._2)
+      maybeRecord = store.find((id, record) => id == recordId && record.publicationState == from).map(_._2)
       count <- maybeRecord
         .map(record =>
           for {
@@ -49,7 +48,7 @@ class CredentialRepositoryInMemory(
     } yield 1
   }
 
-  override def getIssueCredentialRecord(recordId: UUID): Task[Option[IssueCredentialRecord]] = {
+  override def getIssueCredentialRecord(recordId: DidCommID): Task[Option[IssueCredentialRecord]] = {
     for {
       store <- storeRef.get
       record = store.get(recordId)
@@ -63,13 +62,13 @@ class CredentialRepositoryInMemory(
   }
 
   override def updateCredentialRecordProtocolState(
-      recordId: UUID,
+      recordId: DidCommID,
       from: ProtocolState,
       to: ProtocolState
   ): Task[Int] = {
     for {
       store <- storeRef.get
-      maybeRecord = store.find((uuid, record) => uuid == recordId && record.protocolState == from).map(_._2)
+      maybeRecord = store.find((id, record) => id == recordId && record.protocolState == from).map(_._2)
       count <- maybeRecord
         .map(record =>
           for {
@@ -89,7 +88,7 @@ class CredentialRepositoryInMemory(
   }
 
   override def updateWithIssuedRawCredential(
-      recordId: UUID,
+      recordId: DidCommID,
       issue: IssueCredential,
       issuedRawCredential: String,
       protocolState: ProtocolState
@@ -116,7 +115,7 @@ class CredentialRepositoryInMemory(
     } yield count
   }
 
-  override def getValidIssuedCredentials(recordId: Seq[UUID]): Task[Seq[ValidIssuedCredentialRecord]] = {
+  override def getValidIssuedCredentials(recordId: Seq[DidCommID]): Task[Seq[ValidIssuedCredentialRecord]] = {
     for {
       store <- storeRef.get
     } yield store.values
@@ -125,7 +124,7 @@ class CredentialRepositoryInMemory(
       .toSeq
   }
 
-  override def deleteIssueCredentialRecord(recordId: UUID): Task[Int] = {
+  override def deleteIssueCredentialRecord(recordId: DidCommID): Task[Int] = {
     for {
       maybeRecord <- getIssueCredentialRecord(recordId)
       count <- maybeRecord
@@ -139,7 +138,7 @@ class CredentialRepositoryInMemory(
   }
 
   override def updateWithIssueCredential(
-      recordId: UUID,
+      recordId: DidCommID,
       issue: IssueCredential,
       protocolState: ProtocolState
   ): Task[Int] = {
@@ -170,14 +169,14 @@ class CredentialRepositoryInMemory(
     } yield store.values.filter(rec => states.contains(rec.protocolState)).toSeq
   }
 
-  override def getIssueCredentialRecordByThreadId(thid: UUID): Task[Option[IssueCredentialRecord]] = {
+  override def getIssueCredentialRecordByThreadId(thid: DidCommID): Task[Option[IssueCredentialRecord]] = {
     for {
       store <- storeRef.get
     } yield store.values.find(_.thid == thid)
   }
 
   override def updateWithRequestCredential(
-      recordId: UUID,
+      recordId: DidCommID,
       request: RequestCredential,
       protocolState: ProtocolState
   ): Task[Int] = {
@@ -203,11 +202,11 @@ class CredentialRepositoryInMemory(
   }
 
   override def updateCredentialRecordStateAndProofByCredentialIdBulk(
-      idsStatesAndProofs: Seq[(UUID, PublicationState, MerkleInclusionProof)]
+      idsStatesAndProofs: Seq[(DidCommID, PublicationState, MerkleInclusionProof)]
   ): Task[Int] = ???
 
   def updateAfterFail(
-      recordId: UUID,
+      recordId: DidCommID,
       failReason: Option[String]
   ): Task[Int] = for {
     maybeRecord <- getIssueCredentialRecord(recordId)
@@ -237,7 +236,7 @@ object CredentialRepositoryInMemory {
   val maxRetries = 5 // TODO Move to config
   val layer: ULayer[CredentialRepository[Task]] = ZLayer.fromZIO(
     Ref
-      .make(Map.empty[UUID, IssueCredentialRecord])
+      .make(Map.empty[DidCommID, IssueCredentialRecord])
       .map(CredentialRepositoryInMemory(_, maxRetries))
   )
 }